fix:对交叉编译的临时工具链的编译脚本中新增了configure之后的判断逻辑

+++ src/cross_compile_script/file.sh
@@ -6,7 +6,13 @@ pushd build
                --disable-zlib
   make
 popd
+
 ./configure --prefix=/usr --host=$LFS_TGT --build=$(./config.guess)
-make FILE_COMPILE=$(pwd)/build/src/file
-make DESTDIR=$LFS install
+
+if [ "$?" -eq 1 ];
+then
+    exit $?
+fi
+
+make FILE_COMPILE=$(pwd)/build/src/file && make DESTDIR=$LFS install
 rm -v $LFS/usr/lib/libmagic.la
